Updateret filstruktur
