+++ src/KNN.c
@@ -1,4 +1,5 @@
 #include "questionnaire.h"
+#include "data_collection.h"
 #include "KNN.h"
 
 #include <stdio.h>
@@ -9,48 +10,12 @@
 #include <ctype.h>
 #include <limits.h>
 
-#if defined(_WIN32) || defined(_WIN64)
-#include <direct.h>
-#define mkdir(path, mode) _mkdir(path)
-#elif defined(__APPLE__)
-#include <sys/stat.h>
-    #include <unistd.h>
-#else
-    #include <sys/stat.h>
-    #include <sys/types.h>
-#endif
-
 
 char* answers_path;
 char* datast_path = {"Databases/datast.csv"};
 char* result_path;
 
-int make_directory(const char *path) {
-#if defined(_WIN32) || defined(_WIN64)
-    if (_mkdir(path) == 0) {
-        return 0;
-    } else {
-        fprintf(stderr, "Failed to create directory.\n");
-        return -1;
-    }
-#elif defined(__APPLE__)
-    if (mkdir(path, 0777) == 0) {
-            printf("Directory created successfully.\n");
-            return 0;
-        } else {
-            fprintf(stderr, "Failed to create directory.\n");
-            return -1;
-        }
-    #else
-        if (mkdir(path, 0777) == 0) {
-            printf("Directory created successfully.\n");
-            return 0;
-        } else {
-            fprintf(stderr, "Failed to create directory.\n");
-            return -1;
-        }
-#endif
-}
+
 
 void knn() {
