+++ src/data_collection.h
@@ -3,6 +3,8 @@
 #define CPR_MAX_LENGTH 10
 #define HASH_PASSWORD 100
 #define MAX_LINE_LENGTH (USERNAME_MAX_LENGTH + PASSWORD_MAX_LENGTH + CPR_MAX_LENGTH)
+
+
 //Defines a struct for user logins consisting of a username and a password
 typedef struct
 {
@@ -13,7 +15,7 @@ typedef struct
 }Logins;
 
 extern Logins current_user;
-
+int make_directory(const char *path);
 void login();
 Logins load_user();
 char login_or_signup();
